Update all scripts

+++ template/test_rtl/mutation/template/run.sh
@@ -4,22 +4,28 @@
     echo "        [MCY] Sourcing OSS CAD Suite environment..."
     source ~/oss-cad-suite/environment
     if [ $? -ne 0 ]; then
-        echo "        [MCY] Failed to source OSS CAD Suite environment. Exiting script."
+        echo "        [MCY] FAIL: Failed to source OSS CAD Suite environment. Exiting script."
         exit 1
     fi
 
     # Copy original rtl here
-    cp ${PWD}/../../../rtl/average_filter.v .
+    cp ${PWD}/../../../rtl/template.v .
 
-    # Copy testbench here
-    cp ${PWD}/../../simulation/icarus/average_filter/testbench.v .
+    # Append `define MCY after `timescale 1ps/1ps
+    sed '/\`timescale 1ps\/1ps/a \
+    \`define MCY' template.v > template_temp.v
+
+    #replace the orginal testbench file with the temp file.
+    mv template_temp.v template.v
+
+    # Copy template here
+    cp ${PWD}/../../simulation/icarus/template/testbench.v .
 
     # Append `define MCY after `timescale 1ps/1ps
     sed '/\`timescale 1ps\/1ps/a \
     \`define MCY' testbench.v > testbench_temp.v
 
     #replace the orginal testbench file with the temp file.
-    #rm testbench.v
     mv testbench_temp.v testbench.v
 
     # Move create scripts to $SCRIPTS
@@ -29,9 +35,14 @@
     # Generate mutations using mcy
     echo "        [MCY] Generating mutations using mcy..."
     mcy purge; mcy init; mcy run -j8
+    if [ $? -ne 0 ]; then
+        echo "        [MCY] FAIL: mcy process failed. Exiting script."
+        exit 1
+    fi
+    echo "        [MCY] PASS: mcy process passed"
 
     # Remove testbench
     rm testbench.v
 
     # Copy original rtl here
-    rm average_filter.v
+    rm template.v
